move populate function out
